+++ funvas/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.4
+
+* Added the member `s` to `Funvas`, which is a shortcut for the available size (that is also
+  available through `x.width` and `x.height`).
+
 ## 0.1.3+5
 
 * Updated example app to use `flutter_lints`.